Text to speech issue fixed
