+++ app/dashboard/_components/InterviewItemCard.jsx
@@ -17,11 +17,9 @@ function InterviewItemCard({interviewInfo}) {
         <h2 className='font-bold text-primary'>{interviewInfo?.jobPosition}</h2>
         <h2 className='text-sm text-gray-600'>{interviewInfo?.jobExperience} Years of Experience</h2>
         <h2 className='text-xs text-gray-500'>Created At: {interviewInfo.createdAt}</h2>
-        <div className='flex justify-between mt-2 gap-5'>
-           
-            <Button size="sm" variant="outline"  className="w-full" onClick={onFeedback}>Feedback</Button>
-        
-            <Button size="sm" className="w-full" onClick={onStart}>Start</Button>
+        <div className='grid grid-cols-2 mt-2 gap-5'>
+            <Button size="sm" variant="outline"  onClick={onFeedback}>Feedback</Button>
+            <Button size="sm" onClick={onStart}>Start</Button>
         </div>
     </div>
   )
